URL encode ${instanceKeyword}
A semicolon in the title fails the query.

There exists titles that contain a semicolon. Example:

"title" : "Cantatas for bass 4 Ich habe genug : BWV 82 / Johann Sebastian Bach ; Matthias Goerne, baritone ; Freiburger Barockorchester, Gottfried von der Goltz, violin and conductor",

The title is used as ${instanceKeyword} and this query parameter:
query=(keyword all "${instanceKeyword}") sortBy title

If the semicolon is not URL encoded it ends the query parameter,
this results in a CQL parse exception because the trailing " is missing.

Solution:
URL encode ${instanceKeyword}.